Move pass profile to different module
