+++ perun/postprocess/normalizer/run.py
@@ -3,7 +3,8 @@
 import click
 
 import perun.logic.runner as runner
-from perun.utils.helpers import PostprocessStatus, pass_profile
+from perun.utils.helpers import PostprocessStatus
+from perun.profile.factory import pass_profile
 
 __author__ = 'Tomas Fiedor'
