+++ perun/view/raw/run.py
@@ -7,7 +7,8 @@ formatting and visualization techniques at all.
 import click
 import termcolor
 
-from perun.utils.helpers import RAW_ATTRS, RAW_ITEM_COLOUR, RAW_KEY_COLOUR, pass_profile
+from perun.utils.helpers import RAW_ATTRS, RAW_ITEM_COLOUR, RAW_KEY_COLOUR
+from perun.profile.factory import pass_profile
 
 __author__ = 'Tomas Fiedor'
