+++ perun/view/scatter/run.py
@@ -7,7 +7,7 @@ import perun.utils.cli_helpers as cli_helpers
 import perun.utils.bokeh_helpers as bokeh_helpers
 import perun.view.scatter.factory as scatter_factory
 
-from perun.utils.helpers import pass_profile
+from perun.profile.factory import pass_profile
 
 __author__ = 'Jiri Pavela'
